refactor #98 : 재사용 컴포넌트(LeftSidebar), 도메인별 소비 컴포넌트(FeedbackPageNavigator) 이름 분리
